feat: classFee- 0원부터 시작하도록 + 짜잘하게 덜 고친 부분
