+++ src/components/modal/HomeFilter.tsx
@@ -8,7 +8,7 @@ import {
   allLocationList,
   allGenreList,
   allTimeSelect,
-} from '@/data/upload-data';
+} from '@/data/class-data';
 import DuplicationSelect from '@/components/upload/DuplicationSelect';
 import Select from '@/components/upload/Select';
 import ClassDay from '@/components/upload/ClassDay';
@@ -26,7 +26,6 @@ import {
   classFeeState,
   classLevelState,
   classWayState,
-  clickedTimeState,
   genreListState,
   homeFilterValueListState,
   isClickedGenreState,
