+++ src/data/class-data.ts
@@ -34,9 +34,70 @@ export const filterWayList = [
 
 export const classFeeList = [
   { id: 0, name: '전체 가격' },
-  { id: 1, name: '1,000원 ~ 10,000원' },
+  { id: 1, name: '0원 ~ 10,000원' },
   { id: 2, name: '10,000원 ~ 30,000원' },
   { id: 3, name: '30,000원 ~ 60,000원' },
   { id: 4, name: '60,000원 ~ 90,000원' },
   { id: 5, name: '100,000원 이상' },
+];
+
+export const allLocationList = [
+  { id: 0, name: '강남구', isShort: true },
+  { id: 1, name: '강동구', isShort: true },
+  { id: 2, name: '강북구', isShort: true },
+  { id: 3, name: '강서구', isShort: true },
+  { id: 4, name: '관악구', isShort: true },
+  { id: 5, name: '광진구', isShort: true },
+  { id: 6, name: '구로구', isShort: true },
+  { id: 7, name: '금천구', isShort: true },
+  { id: 8, name: '노원구', isShort: true },
+  { id: 9, name: '도봉구', isShort: true },
+  { id: 10, name: '동대문구', isShort: true },
+  { id: 11, name: '동작구', isShort: true },
+  { id: 12, name: '마포구', isShort: true },
+  { id: 13, name: '서대문구', isShort: true },
+  { id: 14, name: '서초구', isShort: true },
+  { id: 15, name: '성동구', isShort: true },
+  { id: 16, name: '성북구', isShort: true },
+  { id: 17, name: '송파구', isShort: true },
+  { id: 18, name: '양천구', isShort: true },
+  { id: 19, name: '영등포구', isShort: true },
+  { id: 20, name: '용산구', isShort: true },
+  { id: 21, name: '은평구', isShort: true },
+  { id: 22, name: '종로구', isShort: true },
+  { id: 23, name: '중구', isShort: true },
+  { id: 24, name: '중량구', isShort: true },
+];
+
+export const allGenreList = [
+  { id: 0, name: '락킹', isShort: true },
+  { id: 1, name: '왁킹', isShort: true },
+  { id: 2, name: '팝핑', isShort: true },
+  { id: 3, name: '재즈', isShort: true },
+  { id: 4, name: '걸스 힙합', isShort: false },
+  { id: 5, name: '힙합', isShort: true },
+  { id: 6, name: '텃팅', isShort: true },
+  { id: 7, name: '디짓', isShort: true },
+  { id: 8, name: '트월크', isShort: false },
+  { id: 9, name: '힐댄스', isShort: false },
+  { id: 10, name: '브레이킹', isShort: false },
+  { id: 11, name: '크럼프', isShort: false },
+  { id: 12, name: '하우스', isShort: false },
+  { id: 13, name: '코레오', isShort: false },
+  { id: 14, name: '컨템', isShort: true },
+  { id: 15, name: '보깅', isShort: true },
+  { id: 16, name: '소울', isShort: true },
+  { id: 17, name: '프리스타일', isShort: false },
+  { id: 18, name: '기타', isShort: true },
+];
+
+export const allTimeSelect = [
+  { id: 0, name: '오전', isShort: true },
+  { id: 1, name: '오후', isShort: true },
+  { id: 2, name: '저녁', isShort: true },
+  { id: 3, name: '새벽', isShort: true },
+  { id: 4, name: '오전-오후', isShort: false },
+  { id: 5, name: '오후-저녁', isShort: false },
+  { id: 6, name: '저녁-새벽', isShort: false },
+  { id: 7, name: '새벽-오전', isShort: false },
 ];
